requirements.txt
numpy==1.15.4
tissueloc==1.5.0
pycontour==1.1.0
Pillow==5.3.0
openslide-python==1.1.1
scikit-image==0.14.1
opencv-python==3.4.4.19
